Make the lint gate cover the engine, and clear what it found
The repo-wide pass planned for after spec 031. Three things, in one sweep
because they are the same question asked of three toolchains.

`verify` excluded the engine crate entirely — bevy is taken with no
windowing backend, so it cannot build for the host, and the conclusion
drawn from that was that the crate could not be linted at all. It can:
wasm-pack already builds it for wasm32-unknown-unknown, so any environment
that can build the engine can lint it. It had eleven warnings nobody had
ever seen. `verify` now has an "engine lint (wasm)" step, and the crate is
clean.

Thirty of the workspace's thirty-three clippy warnings are gone. Most were
mechanical — a deprecated `TOTP` alias and `Nonce::from_slice`, collapsible
ifs, redundant closures, a cast to its own type, items after a test module.
Three were not:

- `spawn_listen_task` leaked its stop-sender with `mem::forget` to keep the
  watch alive. The sender now moves into the delivery task, which is exactly
  the lifetime it needed.
- The heartbeat module carried its own copy of the presence rule —
  `is_connected` and `PRESENCE_TIMEOUT_SECS` — that nothing called. The rule
  that runs is `thunderforge_presence`'s. Removed, along with the test that
  existed only to assert the two copies agreed.
- Two warnings are documented rather than fixed: the `println!`s in a
  reporting benchmark whose output *is* the table, and the `process::exit`
  at boot before the listener is bound.

Web lint went from ten warnings to zero, which matters because `verify`
runs eslint with `--max-warnings=0` — the gate was red. `badgeVariants`,
`buttonVariants` and `tabsListVariants` were exported and imported nowhere,
so they are no longer exported. The theme, game-system and system-hooks
contexts are split from their providers: a module exporting both a
component and a hook cannot fast-refresh, so editing either half forced a
full reload. The three `exhaustive-deps` warnings are real fixes, not
suppressions — `useCanvasEngine` destructures the three option fields it
actually depends on, and `WorldPage`'s two reconnect listeners read
`reconcileOptions` through a ref, because naming it as a dependency would
restart the heartbeat and the live-sync subscription every time the user or
the world store changed.

Three dead-code warnings are deliberately left standing. `drop_for_subject`
and `cancel_for_requester` are written, documented, tested by nothing and
called by nothing — the same shape as the two dead layers found during 031.
Wiring them changes behaviour and belongs in a spec, and silencing them
would hide exactly the finding worth keeping.

Verified: all 7 verify checks pass, 548 server / 359 canvas-core / 327 web
unit tests pass. The three missing server tests are the three deleted with
the duplicated presence rule.

+++ crates/thunderforge-canvas-core/src/grid.rs
@@ -320,7 +320,7 @@ impl GridSpec {
             GridKind::HexPointyTop | GridKind::HexFlatTop => {
                 let (dq, dr) = (a.q - b.q, a.r - b.r);
                 // Cube distance via the implicit third axis s = -q - r.
-                ((dq.abs() + dr.abs() + (dq + dr).abs()) / 2) as i32
+                (dq.abs() + dr.abs() + (dq + dr).abs()) / 2
             }
         }
     }
